User commit

+++ src/main/java/com/example/demo/entity/Student.java
@@ -2,10 +2,10 @@ package com.example.demo.entity;
 
 import java.time.LocalDate;
 public class Student {
-    private int id;
+    private long id;
     private String name;
     private String email;
-    public int getId() {
+    public long getId() {
         return id;
     }
     public void setId(int id) {
@@ -23,7 +23,7 @@ public class Student {
     public void setEmail(String email) {
         this.email = email;
     }
-    public Student(int id, String name, String email) {
+    public Student(long id, String name, String email) {
         this.id = id;
         this.name = name;
         this.email = email;
